[refactor] add fibsem/autofunctions module
Consolidates generic auto-adjustment algorithms into a single coherent module:
- fibsem/autofunctions/gamma.py (from imaging/autogamma.py)
- fibsem/autofunctions/metrics.py (from fm/calibration/metrics.py)
- fibsem/autofunctions/stacking.py (from fm/calibration/focus.py)
- fibsem/autofunctions/integration.py (from fm/calibration/integration.py)
- fibsem/autofunctions/focus.py (auto_focus_beam from calibration.py)

Old import paths preserved via re-export shims for backward compatibility.
FM-specific autofocus entry points (run_autofocus etc.) remain in fm/calibration/.
