+++ fibsem/autofunctions/__init__.py
@@ -1,0 +1,40 @@
+from fibsem.autofunctions.gamma import auto_gamma, apply_gamma, apply_clahe
+from fibsem.autofunctions.focus import auto_focus_beam
+from fibsem.autofunctions.metrics import (
+    laplacian_focus_measure,
+    sobel_focus_measure,
+    variance_focus_measure,
+    tenengrad_focus_measure,
+    get_focus_measure_function,
+    find_best_focus_plane,
+    calculate_focus_quality,
+)
+from fibsem.autofunctions.stacking import (
+    pixel_based_focus_selection,
+    block_based_focus_selection,
+    create_focus_stack_from_selection,
+    create_pixel_based_focus_stack,
+    create_block_based_focus_stack,
+)
+from fibsem.autofunctions.integration import frame_integration, adaptive_frame_integration
+
+__all__ = [
+    "auto_gamma",
+    "apply_gamma",
+    "apply_clahe",
+    "auto_focus_beam",
+    "laplacian_focus_measure",
+    "sobel_focus_measure",
+    "variance_focus_measure",
+    "tenengrad_focus_measure",
+    "get_focus_measure_function",
+    "find_best_focus_plane",
+    "calculate_focus_quality",
+    "pixel_based_focus_selection",
+    "block_based_focus_selection",
+    "create_focus_stack_from_selection",
+    "create_pixel_based_focus_stack",
+    "create_block_based_focus_stack",
+    "frame_integration",
+    "adaptive_frame_integration",
+]
