+++ fibsem/calibration.py
@@ -1,136 +1,17 @@
 import logging
 
 import numpy as np
-import skimage
 
 from fibsem import acquire, utils
 from fibsem import config as cfg
 from fibsem.microscope import FibsemMicroscope
 from fibsem.structures import (
     BeamType,
-    FibsemImage,
-    FibsemRectangle,
     ImageSettings,
     MicroscopeSettings,
     MicroscopeState,
 )
 
-
-def auto_focus_beam(
-    microscope: FibsemMicroscope,
-    settings: MicroscopeSettings,
-    beam_type: BeamType,
-    metric_fn = None, # function to calculate focus metric
-    focus_image_settings: ImageSettings = None,
-    step_size: float = 0.05e-3,
-    num_steps: int = 5,
-    kwargs: dict = {},
-    verbose: bool = False,
-) -> None:
-
-    # TODO: @patrickcleeve2 this could be generalised further if we specify the parameter to sweep through too...
-    # e.g. microscope.set("working_distance", value, beam_type) for auto focus
-    # e.g. microscope.set("beam_current", value, beam_type) for auto stigmation
-    # e.g. microscope.set("beam_current", value, beam_type) for auto beam current
-    # might be too much generalisation though...
-    # also need to specify a default function
-
-    if metric_fn is None:
-        # run the default autofocus routine
-        microscope.auto_focus(beam_type=beam_type)
-        return
-        
-    if focus_image_settings is None:
-        # use preset settings if not defined
-        focus_image_settings = ImageSettings(
-            resolution=[768, 512],
-            dwell_time=200e-9,
-            hfw=100e-6,
-            beam_type=beam_type,
-            save=True,
-            path=settings.image.path,
-            autocontrast=True,
-            autogamma=False,
-            filename=f"{utils.current_timestamp()}_",
-            reduced_area=FibsemRectangle(0.3, 0.3, 0.4, 0.4),
-        )
-
-    # get current working distance
-    current_wd = microscope.get("working_distance", beam_type)
-
-    if verbose:
-        logging.info(f"{metric_fn.__name__} based auto-focus routine")
-        logging.info(f"doc: {metric_fn.__doc__}")
-        logging.info(f"initial working distance: {current_wd:.2e}")
-
-    # define working distance range
-    min_wd = current_wd - (num_steps * step_size / 2)
-    max_wd = current_wd + (num_steps * step_size / 2)
-    wds = np.linspace(min_wd, max_wd, num_steps + 1)
-
-    # loop through working distances and calculate the sharpness (acutance)
-    # highest acutance is best focus
-    metrics = []
-    for i, wd in enumerate(wds):
-
-        logging.info(f"image {i}: {wd:.2e}")
-        microscope.set("working_distance", wd, beam_type)
-
-        focus_image_settings.filename = f"{utils.current_timestamp()}_sharpness_{i}"
-        img = acquire.new_image(microscope, focus_image_settings)
-
-        # calculate focus metric 
-        metric = metric_fn(img, **kwargs)
-        metrics.append(metric)
-
-    # select working distance with highest metric
-    idx = np.argmax(metrics)
-
-    if verbose:
-        pairs = list(zip(wds, metrics))
-        logging.info([f"{wd:.2e}: {metric:.4f}" for wd, metric in pairs])
-        logging.info(f"{idx}, {wds[idx]:.2e}, {metrics[idx]:.4f}")
-
-    # set working distance
-    microscope.set(
-        key="working_distance",
-        value=wds[idx],
-        beam_type=beam_type,
-    )
-
-
-    return
-
-def _sharpness(img:FibsemImage, **kwargs) -> float:
-    """Calculate sharpness (accutance) of an image.
-    (Acutance: https://en.wikipedia.org/wiki/Acutance)
-
-    Args:
-        img (FibsemImage): _description_
-        disk_size (int, optional): _description_. Defaults to 5.
-
-    Returns:
-        _type_: _description_
-    """
-    from skimage.filters.rank import gradient
-    from skimage.morphology import disk
-    disk_size = kwargs.get("disk_size", 5)
-    logging.info(f"calculating sharpness (accutance) of image {img}: {disk_size}")
-    return np.mean(gradient(skimage.filters.median(np.copy(img.data)), disk(disk_size)))
-
-def _dog(img: FibsemImage, **kwargs) -> float:
-    """Calculate difference of gaussian (DoG) of an image.
-
-    Args:
-        img (FibsemImage): _description_
-
-    Returns:
-        _type_: _description_
-    """
-    low = kwargs.get("low", 3)
-    high = kwargs.get("high", 9)
-    from skimage.filters import difference_of_gaussians
-    return np.mean(difference_of_gaussians(np.copy(img.data), low, high))
 
 def auto_charge_neutralisation(
     microscope: FibsemMicroscope,
